clean pam
