+++ libstns/pam_fixtures/auth_01.conf
@@ -1,0 +1,1 @@
+wrapper_path = "./pam_fixtures/response_01"
