+++ libstns/pam_fixtures/auth_05.conf
@@ -1,0 +1,1 @@
+wrapper_path = "./pam_fixtures/response_05"
