--- pam/fixtures/auth_01.conf
@@ -1,1 +1,0 @@
-wrapper_path = "./fixtures/response_01"
