--- pam/fixtures/auth_02.conf
@@ -1,1 +1,0 @@
-wrapper_path = "./fixtures/response_02"
