--- pam/fixtures/auth_03.conf
@@ -1,1 +1,0 @@
-wrapper_path = "./fixtures/response_03"
