--- pam/fixtures/auth_04.conf
@@ -1,1 +1,0 @@
-wrapper_path = "./fixtures/response_04"
