--- pam/fixtures/auth_05.conf
@@ -1,1 +1,0 @@
-wrapper_path = "./fixtures/response_05"
